Moved JsonWriter class to mediaProperty folder
